University of Louisville Project
This project was to create a dynamic site that could be used for both
football or basketball games. The site uses Javascript, jQuery, HTML4,
CSS3 and most of the code was created by myself. This site allows the
current status of our available seats for an event to be populated.